Check player's element on login.
If the player is not the same element as the one they have in the team,
it will boot them from the team. If the player is ownership of said
team, ownership will be transferred to another player on the team. (Goes
in order of Air > Water > Earth > Fire > Chi), checks to see which
element the team has and gives ownership to one of those bending types.
If the team has no more benders, the team is destroyed.

This check automatically bypasses anyone with more than one element.